refactor(lib): extract plugin build-step + script pre-glob into plugin_build (#217)
Move the two build-time helper clusters out of lib.rs into src/plugin_build.rs:

- build-step runner: execute_build_command, collect_build_rtp, run_build_shell,
  unwrap_anonymous_lua_function, run_build_lua, parse_build_command.
- loader script pre-glob: build_plugin_scripts + collect_colorschemes /
  collect_denops_plugins / collect_source_files.

All pub(crate), glob-imported by lib.rs. resolve_plugin_dst and find_lua stay in
lib.rs (widely used elsewhere) and are now pub(crate) so plugin_build can call
them via crate::.

Behaviour-preserving code motion. fmt --check + clippy -D warnings + test (805)
+ --doc all green.
